Fixes a bug in vote.py with symlinked labels failing to be found.
vote.py looked for template labels by searching in /path/to/image.mnc/../labels/image_labels.mnc
but used the canonical path (via os.path.realpath) rather than the given path (which could
be a symlink).  If the label file existed relative to the symlink path but not the canonical
path, then it wouldn't be found, resulting in unexpected errors (for instance, atlases without
labels).

The fix is to use the given path when searching for labels, not the canonical one.