Vanilla implementation of MAGeT brain completed (untested tho')
'mb run' now runs through all registrations, label propagation, and
voting. Either 'parallel' or 'qbatch' can be set as the queueing system,
and then the commands will be appropriately fired off, and in the case
of 'qbatch' the appropriate dependencies will be set.

Additionally, a specific stage can be run by naming it after the 'run'
command.  Currently, 'vanilla' runs everything, 'register' runs
atlas-to-template registrations, and 'vote' runs all subject-specific
stages.  (Perhaps in the future we'll call that 'subject-vote' to
distinguish it purely from the voting stage).

NOTE: at the moment, all files generated except atlas-to-template
transforms and voted labels are placed in a temporary folder. There is
no facility for keeping these 'intermediate files' in the output folder.