Fix for using 'parallel' queue.
This commit is a partial fixup of the logic around queueing up stages when using the 'parallel' queue.

The problem is this: if atlases-template registrations aren't already done, then during construction of the voting stage commands, xfmjoin mistakenly uses a path to the temporary folder for atlas-template xfms.  It should look at the output/registrations folder, but since the registrations haven't been completed at the time of construction, it defaults to the temp folder. This is never a problem when using the vanilla mode (atlas-template registration + voting) with the qbatch queue because the construction of voting commands happens in a job that is run only after the atlas-template commands have run.  When using the 'parallel' queue all of the stages are constructed in a single pass and so this is why the XFMs aren't found.

So, with this comment, I've reworked things so that in order to run using the 'parallel' queue, one needs run the 'register' stage, and the 'voting' stage separately. e.g.

   mb run -q parallel register
   mb run -q parallel voting

That is, the 'vanilla' mode does not work for the parallel queue (a warning is displayed and its treated as only the 'register' stage).

This is obviously a bit of a workaround until we can do something more sensible with handling temporary folders and pipeline construction.